requirements.txt
openai
anthropic
pandas
numpy
python-dotenv
openpyxl
tqdm